Bug 978700: Fix warnings about comparison of values with different signs, r=roc
UltraBlame original commit: d535393b0f71e4d9c1089ca0c7d93e60af033d52

+++ widget/gonk/nsAppShell.cpp
@@ -317,7 +317,7 @@ KeyEventDispatcher::KeyEventDispatcher(const UserInputData& aData,
 {
     
     
-    mDOMKeyCode = (mData.key.keyCode < ArrayLength(kKeyMapping)) ?
+    mDOMKeyCode = (mData.key.keyCode < (ssize_t)ArrayLength(kKeyMapping)) ?
         kKeyMapping[mData.key.keyCode] : 0;
     mDOMKeyNameIndex = GetKeyNameIndex(mData.key.keyCode);
